pyproject.toml
[project]
name = "google-cloud-mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-container>=2.56.1",
    "google-cloud-monitoring>=2.27.1",
    "google-api-python-client>=2.125.0",
    "mcp[cli]>=1.9.0",
    "google-cloud-compute>=1.30.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
